Added processingMode to UI
