+++ src/main/java/de/uniwue/controller/RecognitionController.java
@@ -40,7 +40,8 @@ public class RecognitionController {
         if (recognitionHelper == null) {
             recognitionHelper = new RecognitionHelper(
                 session.getAttribute("projectDir").toString(),
-                session.getAttribute("imageType").toString()
+                session.getAttribute("imageType").toString(),
+                session.getAttribute("processingMode").toString()
             );
             session.setAttribute("recognitionHelper", recognitionHelper);
         }
