+++ src/main/java/de/uniwue/controller/SegmentationDummyController.java
@@ -40,7 +40,8 @@ public class SegmentationDummyController {
         if (segmentationDummyHelper == null) {
             segmentationDummyHelper = new SegmentationDummyHelper(
                 session.getAttribute("projectDir").toString(),
-                session.getAttribute("imageType").toString()
+                session.getAttribute("imageType").toString(),
+                session.getAttribute("processingMode").toString()
             );
             session.setAttribute("segmentationDummyHelper", segmentationDummyHelper);
         }
