+++ src/main/java/de/uniwue/feature/ProcessStateCollector.java
@@ -1,6 +1,12 @@
 package de.uniwue.feature;
 
 import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.io.UnsupportedEncodingException;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import org.apache.commons.io.FilenameUtils;
 
@@ -22,14 +28,22 @@ public class ProcessStateCollector {
     private String imageType;
 
     /**
+     * Processing structure of the project
+     * Possible values: { Directory, Pagexml }
+     */
+    private String processingMode;
+
+    /**
      * Constructor
      *
      * @param projConf Project configuration object
+     * @param projectImageType Type of the project (gray, binary)
      * @param imageType Image type of the project
      */
-    public ProcessStateCollector(ProjectConfiguration projConf, String imageType) {
+    public ProcessStateCollector(ProjectConfiguration projConf, String imageType, String processingMode) {
         this.projConf = projConf;
         this.imageType = imageType;
+        this.processingMode = processingMode;
     }
 
     /**
@@ -111,6 +125,39 @@ public class ProcessStateCollector {
     }
 
     /**
+     * Checks whether the required image files in "OCR/Pages/{pageId}" directory exist or not
+     *
+     * @param pageID Identifier of the page (e.g 0002,0003)
+     * @return Information if the required image files exist
+     */
+    public boolean existLines(String pageId) {
+        File pageXML = new File(projConf.OCR_DIR + pageId + projConf.CONF_EXT);
+        if (!pageXML.exists())
+            return false;
+
+        // Easy and fast check if an end tag of a TextLine exists 
+        // Does not check if the xml is valid
+        try {
+			FileInputStream fis = new FileInputStream(pageXML);
+			byte[] data = new byte[(int) pageXML.length()];
+			fis.read(data);
+			fis.close();
+			String pageXMLContent = new String(data, "UTF-8");
+			
+			Pattern p = Pattern.compile("\\</TextLine\\>");
+			Matcher matcher = p.matcher(pageXMLContent);
+
+			return matcher.find();
+		} catch (UnsupportedEncodingException e) {
+			return false;
+		} catch (FileNotFoundException e) {
+			return false;
+		} catch (IOException e) {
+			return false;
+		}
+    }
+    
+    /**
      * Determines the "RegionExtraction" process state of a given page
      *
      * @param pageID Identifier of the page (e.g 0002,0003)
@@ -128,9 +175,14 @@ public class ProcessStateCollector {
      * @return "LineSegmentation" state of the page
      */
     public boolean lineSegmentationState(String pageId) {
-        // Check that for every region file a ".bin.png" and ".nrm.png exists (successful region extraction)
-        // Check that for every region file a ".pseg.png" exists (successful line segmentation)
-        return existPageDirImageFiles(pageId, true);
+    	if(processingMode.equals("Directory")) {
+			// Check that for every region file a ".bin.png" and ".nrm.png exists (successful region extraction)
+			// Check that for every region file a ".pseg.png" exists (successful line segmentation)
+			return existPageDirImageFiles(pageId, true);
+    	} else {
+    		// Check whether lines in the PageXML files exist
+    		return existLines(pageId);
+    	}
     }
 
     /**
