+++ src/main/java/de/uniwue/helper/ProcessFlowHelper.java
@@ -24,14 +24,22 @@ public class ProcessFlowHelper {
     private String imageType;
 
     /**
+     * Processing structure of the project
+     * Possible values: { Directory, Pagexml }
+     */
+    private String processingMode;
+
+    /**
      * Constructor
      *
      * @param projectDir Absolute path to the project
      * @param imageType Image type of the project
+     * @param processingMode Processing structure of the project (Directory, Pagexml)
      */
-    public ProcessFlowHelper(String projectDir, String imageType) {
+    public ProcessFlowHelper(String projectDir, String imageType, String processingMode) {
         this.projConf = new ProjectConfiguration(projectDir);
         this.imageType  = imageType;
+        this.processingMode = processingMode;
     }
 
     /**
@@ -43,7 +51,7 @@ public class ProcessFlowHelper {
      */
     public String[] getValidPageIds(String[] initialPageIds, String checkProcess) {
         // Needed to determine process states
-        ProcessStateCollector procStateCol = new ProcessStateCollector(projConf, imageType);
+        ProcessStateCollector procStateCol = new ProcessStateCollector(projConf, imageType, processingMode);
 
         // Verify state of each page for the given process and store the successful ones
         Set<String> processedPages = new TreeSet<String>();
