+++ src/main/java/de/uniwue/helper/ResultGenerationHelper.java
@@ -74,11 +74,12 @@ public class ResultGenerationHelper {
      *
      * @param projectDir       Path to the project directory
      * @param projectImageType Type of the project (binary, gray)
+     * @param processingMode Processing structure of the project (Directory, Pagexml)
      */
-    public ResultGenerationHelper(String projectDir, String projectImageType) {
+    public ResultGenerationHelper(String projectDir, String projectImageType, String processingMode) {
         projConf = new ProjectConfiguration(projectDir);
         processHandler = new ProcessHandler();
-        procStateCol = new ProcessStateCollector(projConf, projectImageType);
+        procStateCol = new ProcessStateCollector(projConf, projectImageType, processingMode);
         genericHelper = new GenericHelper(projConf);
     }
